==> Notebooks/01_Core_preprocessing.Rmd <==
---
title: "Preprocessing"
output: html_notebook
---



```{r - load libraries }
library(dplyr)    
library(tidyr)    
library(forcats) 
library(here)

```

```{r - source }
# Source general and dataset-specific cleaning scripts

source(here("Scripts/Cleaning_utils.R"))
source(here("Scripts/Df_specific_utils.R"))
source(here("Scripts/in24_fg_utils.R"))
source(here("Scripts/diet_specific_utils.R"))

```

```{r - data }
#load data
df <- read.delim(here("Data/Raw/Assessment_masterfile.txt"))
FFQ_original <- read.delim(here("Data/Raw/FFQ_results.19062024.txt"))
in24_t1 <- read.delim(here("Data/Raw/intake24_t1.23052024.txt"))
```


# Assessment masterfile
```{r - preprocess df }
# Clean df_t1
# dataset-specific preprocessing function to df_t1
df_list <- preprocess_data(df, 1)

# Extract df
df_t1 <- df_list$data_all 
#we are only working on df_t1_p but we need this for preprocessing of FFQ and be sure that included pp match regardless in person or total in case we need tot 
df_p_t1 <- df_list$data_p

# Preview the cleaned data
head(df_t1)
head(df_p_t1)
```


# FFQ - clean, preprocess, adjust
```{r - preprocess FFQ }
#clean FFQ
# Apply the dataset-specific preprocessing function to the FFQ dataset
FFQ_t1_list <- preprocess_FFQ(FFQ_original, df_t1, df_p_t1, "T2")

FFQ_p_ni_t1 <- FFQ_t1_list$FFQ_p

# Preview the cleaned FFQ data
head(FFQ_p_ni_t1)
```


# in24 - clean and preprocess
```{r - preprocess intake24 }
#clean In24
head(in24_t1)
#apply the function
in24_list <- preprocess_in24(in24_t1, df_p_t1$ID)

# Extract nutrient, food, and food group intake data frames
#in24_fi_ni_t1 <- in24_list$food_nutrient_intake  #fi and ni 
in24_ni_t1 <- in24_list$nutrient_intake          #ni only
in24_fi_t1 <- in24_list$food_intake              #fi only
in24_fg_t1 <- in24_list$food_group_intake        #food groups
```
