==> Notebooks/21_GAMs_conIV.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
 library(purrr)
library(tidyr)
library(ggplot2)
library(ggbeeswarm)
library(mia)
library(dplyr)
library(openxlsx)
library(mgcv)
library(gratia)
library(tibble)
library(kableExtra)
library(here)

```

```{r - upload dependencies }
source(here("Scripts/Gm_utils.R"))

```

```{r - upload data }
tse_taxa_ab20 <- readRDS(here("Data/Processed/tse_taxa_ab20.rds"))
metadata_t1<- readRDS(here("Data/Processed/metadata_t1.rds"))
tse_gbm <- readRDS(here("Data/Processed/tse_gbm.rds"))

```

```{r gbm model }

# subset based on N > 20 & preprocess
sum(rowData(tse_gbm)$N_presence < 20) # include all
df_gbm <- assay(tse_gbm, "standardize")

# define l6 feats
feats_gbm <- rownames(df_gbm)

# make names 
feats_gbm <- make.names(feats_gbm)
rownames(df_gbm) <- make.names(rownames(df_gbm))

# preprocess 
df_gbm <- df_gbm %>% t() %>% as.data.frame() %>% rownames_to_column("ID") %>% left_join(metadata_t1[, c("ID", "HEI_cluster_2", "STAI.s_sc_res")], by = "ID")

# zero df based on count
tse_count <- tse_gbm[rowData(tse_gbm)$N_presence >= 20]
rownames(tse_count) <-make.names(rownames(tse_count))
tse_count <- assay(tse_count) %>% t() %>% as.data.frame() %>% rownames_to_column("ID")

identical(rownames(df_gbm), rownames(tse_count))


# apply lmrob through each feature including only IDs with feature > 0 

# initialize
gam_gbm <- setNames(vector("list", length(feats_gbm)), feats_gbm)

stopifnot(identical(df_gbm$ID, tse_count$ID))

# loop
for (feature in feats_gbm) {
  keep <- tse_count[[feature]] > 0  # or >= 0, or add !is.na(...)
  df   <- df_gbm[keep, , drop = FALSE]

  fmla  <- form <- as.formula(paste0(
    "STAI.s_sc_res ~ s(", feature, ", k = 4)"))

  model <- gam(fmla, data = df, family = scat(link = "identity"), method = "REML" )

  gam_gbm[[feature]] <- list(model = model, N = sum(keep))
}

# extract models
gam_gbm_models <- lapply(gam_gbm, function(x) x$model)

### 3. table
gam_gbm_df <- gam_table(gam_gbm_models, length(feats_gbm))
gam_gbm_df <- gam_gbm_df %>% as.data.frame() %>% rownames_to_column("Feature")

# extract Ns
Ns <- sapply(gam_gbm, function(x) x$N) %>% as.data.frame() %>% rownames_to_column("Feature")
colnames(Ns)[2] <- "N"

# merge 
gam_gbm_df <- left_join(gam_gbm_df, Ns, by = "Feature")

# rownames
rownames(gam_gbm_df) <- gam_gbm_df$Feature
rownames(gam_gbm_df) <- gsub("[._]", " ", rownames(gam_gbm_df))

gam_gbm_df <- gam_gbm_df[, -1]

# save excel
write.xlsx(gam_gbm_df, here("Excel files/gam_gbm_cont.xlsx"), rowNames = T)

```

```{r - gbm + hei }

# initialize
gam_gbm.hei <- setNames(vector("list", length(feats_gbm)), feats_gbm)

stopifnot(identical(df_gbm$ID, tse_count$ID))

# loop
for (feature in feats_gbm) {
  keep <- tse_count[[feature]] > 0  # or >= 0, or add !is.na(...)
  df   <- df_gbm[keep, , drop = FALSE]

  fmla  <- form <- as.formula(paste0(
    "STAI.s_sc_res ~ s(", feature, ", k = 4) + HEI_cluster_2"))

  model <- gam(fmla, data = df, family = scat(link = "identity"), method = "REML" )

  gam_gbm.hei[[feature]] <- list(model = model, N = sum(keep))
}

# extract models
gam_gbm.hei_models <- lapply(gam_gbm.hei, function(x) x$model)

### 3. table
gam_gbm.hei_df <- gam_table(gam_gbm.hei_models, length(feats_gbm))
gam_gbm.hei_df <- gam_gbm.hei_df %>% as.data.frame() %>% rownames_to_column("Feature")

# extract Ns
Ns <- sapply(gam_gbm.hei, function(x) x$N) %>% as.data.frame() %>% rownames_to_column("Feature")
colnames(Ns)[2] <- "N"

# merge 
gam_gbm.hei_df <- left_join(gam_gbm.hei_df, Ns, by = "Feature")

# rownames
rownames(gam_gbm.hei_df) <- gam_gbm.hei_df$Feature
rownames(gam_gbm.hei_df) <- gsub("[._]", " ", rownames(gam_gbm.hei_df))

gam_gbm.hei_df <- gam_gbm.hei_df[, -1]

# save excel
write.xlsx(gam_gbm.hei_df, here("Excel files/gam_gbn.hei_cont.xlsx", rowNames = T))

```


```{r - visualize inositol }

gam_adj <- gam(
  STAI.s_sc_res ~ s(scale(`MGB037...Inositol.synthesis`), k = 4) + HEI_cluster_2, 
  data = df_gbm,
  family = scat(link = "identity"),
  method = "REML" 
)


p <- draw(gam_adj, select = 1) +
    labs(x = "Clr- abundance", y = "Residualsed STAI-s", title = "Inositol synthesis") +
    geom_point(
      data = df_gbm,
      aes(x = !!sym("MGB037...Inositol.synthesis"), y = STAI.s_sc_res, color = HEI_cluster_2),
      inherit.aes = FALSE,
      alpha = 0.7,
      size = 4, cex = 3
    ) +
  geom_rug(
    data = df_gbm,
    aes(x = !!sym("MGB037...Inositol.synthesis"), color = HEI_cluster_2),
    sides = "b", # bottom rug
    inherit.aes = FALSE,
    alpha = 0.8
  ) +
  scale_color_manual(
    name = "HEI cluster",
    values = c("1" = "#1b9e77", "2" = "#d95f02")) +
    theme_minimal() +
  annotate(
    "text",
    x = Inf, y = Inf, 
    label = "edf = 2.85, R² = 0.29, p = 0.000, q = 0.000",
    hjust = 1.3, vjust = 1.5, size = 7
  ) +
 theme(
      axis.title = element_text(size = 18, face = "bold"),
      axis.text = element_text(size = 13, face = "bold"),
      plot.title = element_text(size = 19, face = "bold",  hjust = 0.5),
      legend.title = element_text(size = 13, face = "bold"),
      legend.text = element_text(size = 12)
    ) 


ggsave(here("Plots/plot_inositol.png"), bg = "white", plot = p
       , width = 10, height = 7, dpi = 300)

```
