==> Notebooks/22_clustering_heatmap.Rmd <==
---
title: "Gm taxa - GBMs hierarchical clustering"
output: html_notebook
---

```{r - libraries }
library(tibble)
library(mia)
library(dplyr)
library(robustbase)
library(kableExtra)
library(purrr)
library(ComplexHeatmap)
library(here)

```

```{r - upload dependencies }
source(here("Scripts/Gm_utils.R"))

```


```{r - load data }
tse_taxa_sub30 <- readRDS(here("Data/Processed/tse_taxa_sub30.rds"))
tse_taxa_ab20 <- readRDS(here("Data/Processed/tse_taxa_ab20.rds"))
tse_gbm <- readRDS(here("Data/Processed/tse_gbm.rds"))
tse_gbm_count <- readRDS(here("Data/Processed/tse_gbm_count.rds"))
tse_l6_count <- readRDS(here("Data/Processed/tse_l6_count.rds"))

```

```{r - preprocess }

# binary x
x_bin <- assay(tse_taxa_sub30, "binarized")[c("Ruminococcus_B gnavus", "Flavonifractor plautii", "Faecalibacterium prausnitzii_J", "Faecalibacterium prausnitzii_E"), ]  %>% t() %>% as.data.frame() %>% rownames_to_column("ID")
# con x
x_con <- assay(tse_taxa_ab20, "clr")[c("Bilophila wadsworthia", "unclassified_31", "Bacteroides thetaiotaomicron", "unclassified_12", "Fusicatenibacter saccharivorans", 
                                       "Roseburia intestinalis", "Bacteroides caccae", "Parabacteroides distasonis", "unclassified_34"), ] %>% t() %>% as.data.frame() %>% rownames_to_column("ID")
# y
y_gbm <- assay(tse_gbm, "standardize")[c("MGB037 - Inositol synthesis", "MGB021 - GABA synthesis II", "MGB053 - Butyrate synthesis II", "MGB054 - Propionate synthesis II"), ] %>% t() %>% as.data.frame() %>% rownames_to_column("ID")

# y clr non std 
y_gbm_clr <- assay(tse_gbm, "clr")[c("MGB037 - Inositol synthesis", "MGB021 - GABA synthesis II", "MGB053 - Butyrate synthesis II", "MGB054 - Propionate synthesis II"), ] %>% t() %>% as.data.frame() %>% rownames_to_column("ID")

# HEI
hei <- colData(tse_taxa_sub30)[, "HEI_cluster_2", drop = F] %>% as.data.frame() %>% rownames_to_column("ID") 

```

```{r - CTRL }

# set control parameters for the S- and M- steps

ctrl <- lmrob.control(
  setting = "KS2014",
  maxit.scale = 500, 
  k.max      = 2000, 
  max.it     = 1000)

```


```{r - binary table }

# bin models
bin_df <- dplyr::left_join(x_bin, y_gbm, by = "ID") %>% dplyr::left_join(hei, by = "ID")

# define vars
iv <- colnames(x_bin)[-1]
dv <- colnames(y_gbm)[-1]

# run models 
models_bin <- lapply(dv, function(y) {
  mods <- lapply(iv, function(x) {
    frml <- as.formula(paste0("`", y , "` ~ `", x, "` + HEI_cluster_2"))
    lmrob(frml, data = bin_df, control = ctrl)
  })
  names(mods) <- iv
  table <- lmrob_table(mods, length(iv))
})
# names
names(models_bin) <- dv

# beta list & df
beta_bin_list <- lapply(seq_along(models_bin), function(i) {
  x <- models_bin[[i]]
  beta_df <- x[, "std_beta", drop = F] %>% rownames_to_column("Specie")
  colnames(beta_df)[2] <- names(models_bin)[[i]]
  beta_df
  }
  )

beta_bin <- purrr::reduce(beta_bin_list, dplyr::full_join, by = "Specie")

# qval list and df
qval_bin_list <- lapply(seq_along(models_bin), function(i) {
  x <- models_bin[[i]]
  qval_df <- x[, "qval", drop = F] %>% rownames_to_column("Specie")
  colnames(qval_df)[2] <- names(models_bin)[[i]]
  qval_df
  }
  )

qval_bin <- purrr::reduce(qval_bin_list, dplyr::full_join, by = "Specie")

        
```

```{r - run models }

# con models 
con_df <- dplyr::left_join(x_con, y_gbm, by = "ID")  %>% dplyr::left_join(hei, by = "ID")
colnames(con_df) <- make.names(colnames(con_df))

# define vars
iv <- make.names(names(x_con)[-1])
dv <- make.names(names(y_gbm_clr)[-1])

# IDs aligned check
stopifnot(
  identical(con_df$ID, tse_l6_count$ID),
  identical(con_df$ID, tse_gbm_count$ID)
)

# models
models_con <- lapply(dv, function(y) {
  mods <- lapply(iv, function(x) {
    keep <- !is.na(tse_l6_count[[x]]) & !is.na(tse_gbm_count[[y]]) &
            tse_l6_count[[x]] > 0 & tse_gbm_count[[y]] > 0
    
    df <- data.frame(
      yy = as.numeric(scale(con_df[[y]][keep])),
      xx = as.numeric(scale(con_df[[x]][keep])),
      HEI_cluster_2 = con_df$HEI_cluster_2[keep]
    )

    lmrob(yy ~ xx + HEI_cluster_2, data = df, control = ctrl)
  })
  names(mods) <- iv
  lmrob_table(mods, length(iv))
})
# names
names(models_con) <- dv

# beta list & df
beta_con_list <- lapply(seq_along(models_con), function(i) {
  x <- models_con[[i]]
  beta_df <- x[, "std_beta", drop = F] %>% rownames_to_column("Specie")
  colnames(beta_df)[2] <- names(models_con)[[i]]
  beta_df
  }
  )

beta_con <- purrr::reduce(beta_con_list, dplyr::full_join, by = "Specie")

# qval list & df
qval_con_list <- lapply(seq_along(models_con), function(i) {
  x <- models_con[[i]]
  qval_df <- x[, "qval", drop = F] %>% rownames_to_column("Specie")
  colnames(qval_df)[2] <- names(models_con)[[i]]
  qval_df
  }
  )

qval_con <- purrr::reduce(qval_con_list, dplyr::full_join, by = "Specie")

# fix colnames
colnames(beta_con) <- colnames(beta_bin)
colnames(qval_con) <- colnames(qval_bin)
    
```

```{r - bind, stars mx, annotations }

# bind beta
identical(colnames(beta_bin), colnames(beta_con))
beta_df <- rbind(beta_bin, beta_con)
# as matrix
rownames(beta_df) <- beta_df$Specie
beta_mx <- beta_df[, -1] %>% as.matrix()
# fix names
rownames(beta_mx) <- gsub("[._]", " ", rownames(beta_mx))
colnames(beta_mx) <- c("Inositol synthesis", "GABA synthesis II", "Butyrate synthesis II", "Propionate sythesis II")

# bind qval
identical(colnames(qval_bin), colnames(qval_con))
qval_df <- rbind(qval_bin, qval_con)
# as matrix
rownames(qval_df) <- qval_df$Specie
qval_mx <- qval_df[, -1] %>% as.matrix()
# fix names
rownames(qval_mx) <- gsub("[._]", " ", rownames(qval_mx))
colnames(qval_mx) <- c("Inositol synthesis", "GABA synthesis II", "Butyrate synthesis II", "Propionate sythesis II")

identical(rownames(qval_mx), rownames(beta_mx))
identical(colnames(qval_mx), colnames(beta_mx))

# stars mx, q thresholds: <0.05 "***", <0.10 "**", <0.20 "*", else "."
stars_mx <- ifelse(is.na(qval_mx), "",
            ifelse(qval_mx < 0.05, "***",
            ifelse(qval_mx < 0.10, "**",
            ifelse(qval_mx < 0.20, "*", ""))))
dimnames(stars_mx) <- dimnames(qval_mx)

# ann
taxa_ann <- data.frame(
  Specie = (rownames(beta_mx)),
  Metadata = c(rep("Binary", 4), rep ("Continuous", 9)),
  Phylum = c(rep("Bacillota", 4), rep("Bacteroidota", 4), "Actinobacteria", rep("Bacillota", 3), "Desulfobacterota")
)

rownames(taxa_ann) <- taxa_ann$Specie
taxa_ann$Metadata <- factor(taxa_ann$Metadata)
taxa_ann$Phylum <- factor(taxa_ann$Phylum)
taxa_ann <- taxa_ann[rownames(beta_mx), c("Metadata", "Phylum") , drop = FALSE] 



```


# HIERARCHICAL CLUSTERING
# clustering distance: 
# euclidian: gives weight to magnitude and direction
# corr: gives weight to direction --> cluster taxa by directionality patterns
## pearson capture linear similarities between beta coeff, spearman focus on rank orders 

# Method:
# average: merges clusters whose members are similar on average (mean pairwise distance)
# complete: merges clusters only when all pairs are reasonably close (uses the max pairwise distance). One strong mismatch (e.g., opposite β on a GBM) keeps clusters apart; yields tighter, more conservative clusters and penalizes outliers.

```{r - hierarchical clustering - heatmap }

png(here("Plots/beta_heatmap.png"), width = 2400, height =1800, res = 300)

Heatmap(
  beta_mx,
  name = "Std. β",
  
  row_title = "Specie",
  column_title = "GBM",
  
  clustering_distance_columns = "pearson",
  clustering_method_rows   = "complete",
  clustering_distance_rows = "pearson",
  clustering_method_columns   = "complete",
  
  column_names_rot = 45,
  width  = grid::unit(ncol(beta_mx) * 15, "mm"),
  
  col = circlize::colorRamp2(
    c(-1, 0, 1),           # adjust to your beta value range
    c("#67a9cf", "white", "#ef8a62")),
  
  right_annotation = rowAnnotation(
    df = taxa_ann,
    col = list(
      Metadata = c("Binary" = "#4D4D4D", "Continuous" = "#C0C0C0"),
      Phylum = c("Bacillota"   = "#7CCD7C", "Bacteroidota" = "#FFA54F", "Actinobacteria" = "#EE3B3B", "Desulfobacterota" = "#EE799F")),
    width = unit(4, "mm"),
    annotation_name_gp = grid::gpar(col = NA)
    ),
  
  cell_fun = function(j, i, x, y, w, h, fill) {
  lab <- stars_mx[i, j]
  if (!is.na(lab) && nzchar(lab)) {
    fs <- if (lab == "***") 11 else if (lab == "**") 10 else if (lab == "*") 9 else 8
    grid::grid.text(lab, x = x, y = y,
      gp = grid::gpar(fontsize = fs, fontface = "bold", col = "black"))
  }
  }
)

dev.off()

```
